apps/backend/pyproject.toml
[project]
name = "lifescribe"
version = "0.1.0"
description = "LifeScribe Vault backend"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.8",
  "python-frontmatter>=1.1",
  "pyyaml>=6",
  "httpx>=0.27",
  "puremagic>=1.27",
  "charset-normalizer>=3.3",
  "trafilatura>=1.12",
  "pypdfium2>=4.30",
  "pdfplumber>=0.11",
  "python-docx>=1.1",
  "openpyxl>=3.1",
  "Pillow>=10.4",
  "keyring>=24",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.23",
  "ruff>=0.6",
  "mypy>=1.11",
  "types-PyYAML>=6",
  "pyinstaller>=6.10",
  "keyrings.alt>=5",
  "pytest-httpx>=0.30",
]

[project.scripts]
lifescribe-backend = "lifescribe.api.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lifescribe"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
